+++ ezrules/frontend/src/app/app.routes.ts
@@ -10,6 +10,7 @@ import { UserListsComponent } from './user-lists/user-lists.component';
 import { DashboardComponent } from './dashboard/dashboard.component';
 import { UserManagementComponent } from './user-management/user-management.component';
 import { RoleManagementComponent } from './role-management/role-management.component';
+import { RolePermissionsComponent } from './role-permissions/role-permissions.component';
 
 export const routes: Routes = [
   { path: '', component: DashboardComponent },
@@ -24,5 +25,6 @@ export const routes: Routes = [
   { path: 'user-lists', component: UserListsComponent },
   { path: 'label_analytics', component: LabelAnalyticsComponent },
   { path: 'management/users', component: UserManagementComponent },
-  { path: 'role_management', component: RoleManagementComponent }
+  { path: 'role_management', component: RoleManagementComponent },
+  { path: 'role_management/:id/permissions', component: RolePermissionsComponent }
 ];